requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cloudinary==1.24.0
dj-database-url==0.5.0
Django==3.1.7
django-crispy-forms==1.11.1
django-phone-field==1.8.1
django-widget-tweaks==1.4.8
gunicorn==20.0.4
Pillow==8.1.0
psycopg2==2.8.6
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
whitenoise==5.2.0
